[kineto] Update Kineto submodule
Summary:
Update Kineto submodule to fix an invalid json bug, also update and
move profiler json tracing unit test

Test Plan:
python test/test_profiler.py -v
