+++ test/test_profiler.py
@@ -564,6 +564,32 @@ class TestProfiler(TestCase):
                 assert "test_key2" in metadata
                 assert metadata["test_key2"] == "test_value2"
 
+    def test_profiler_tracing(self):
+        with _profile(use_kineto=kineto_available()) as prof:
+            t1, t2 = torch.ones(1), torch.ones(1)
+            torch.add(t1, t2)
+
+        with TemporaryFileName(mode="w+") as fname:
+            prof.export_chrome_trace(fname)
+            # read the trace and expect valid json
+            # if the JSON generated by export_chrome_trace is not valid, this will throw and fail the test.
+            with io.open(fname, 'r') as f:
+                json.load(f)
+
+        # Same test but for cuda.
+        if not torch.cuda.is_available():
+            return
+
+        device = torch.device("cuda:0")
+        with _profile(use_cuda=True, use_kineto=kineto_available()) as prof:
+            t1, t2 = torch.ones(1, device=device), torch.ones(1, device=device)
+            torch.add(t1, t2)
+
+        with TemporaryFileName(mode="w+") as fname:
+            prof.export_chrome_trace(fname)
+            # Now validate the json
+            with io.open(fname, 'r') as f:
+                json.load(f)
 
 if __name__ == '__main__':
     run_tests()
